update : chia các thư mục 2

+++ backend/src/modules/projectMember/services/projectMember.service.js
@@ -1,21 +1,17 @@
-const Project = require('../models/project.model'); 
+const Project = require('../../project/models/project.model');
 const ProjectMember = require('../../projectMember/models/projectMember.model'); 
+
 const Board = require('../../board/models/board.model'); 
-const Role = require('../../rbac/models/role.model'); // Import thêm Role
-const { Roles, Scopes } = require('../../rbac/constants/rbac.enum'); // Import hằng số
+const Role = require('../../rbac/models/role.model'); 
+const { Roles, Scopes } = require('../../rbac/constants/rbac.enum'); 
 const AppError = require('../../../common/exceptions/AppError'); 
 
 exports.createProject = async (ownerId, projectData) => {
-    // 1. Tạo Project mới[cite: 6]
     const project = await Project.create({ ...projectData, owner_id: ownerId }); 
     
-    // 2. Tìm động ID của chức danh PM trong DB
     const adminRole = await Role.findOne({ name: Roles.PM, scope: Scopes.PROJECT }).lean();
-    if (!adminRole) {
-        throw new AppError('The PM role is not yet present. Please run the RBAC seed file!', 500);
-    }
+    if (!adminRole) throw new AppError('The PM role is not yet present. Please run the RBAC seed file!', 500);
 
-    // 3. Lưu role_id thay vì lưu chữ 'MANAGER'[cite: 6]
     await ProjectMember.create({ 
         project_id: project._id, 
         user_id: ownerId, 
@@ -39,7 +35,7 @@ exports.getProjectDetail = async (projectId) => {
         Board.find({ project_id: projectId }).select('_id name description created_at').lean(), 
         ProjectMember.find({ project_id: projectId }) 
             .populate('user_id', 'full_name email avatar_url') 
-            .populate('role_id', 'name') // Lấy tên Role hiển thị ra ngoài cho Frontend
+            .populate('role_id', 'name') 
             .lean()
     ]);
 
@@ -55,10 +51,12 @@ exports.updateProject = async (projectId, updateData) => {
 };
 
 exports.deleteProject = async (projectId) => {
-    await Promise.all([ 
-        Project.findByIdAndDelete(projectId), 
-        ProjectMember.deleteMany({ project_id: projectId }), 
-        Board.deleteMany({ project_id: projectId })  
-    ]);
-    return true; 
+    const project = await Project.findByIdAndDelete(projectId);
+    if (!project) throw new AppError('Project not found', 404, 'NOT_FOUND');
+    
+    // Khi xóa Project, xóa luôn các Member và Board liên quan
+    await ProjectMember.deleteMany({ project_id: projectId });
+    await Board.deleteMany({ project_id: projectId });
+    
+    return project;
 };
